requirements.txt
python-telegram-bot==12.7
python-aternos==1.0.5